+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/management/ui/template/UITemplateInitializer.java
@@ -72,6 +72,7 @@ public class UITemplateInitializer {
         Layer.KAFKA.name(),
         Layer.PULSAR.name(),
         Layer.BOOKKEEPER.name(),
+        Layer.NGINX.name(),
         "custom"
     };
     private final UITemplateManagementService uiTemplateManagementService;